backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
pandas==2.2.2
openpyxl==3.1.2
groq==0.9.0
networkx==3.3
pydantic==2.7.1
python-multipart==0.0.9
aiofiles==23.2.1